requirements.txt
pymongo
requests
Appium-Python-Client
selenium
redis